pyproject.toml
[project]
name = "scholarmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "arxiv>=2.3.1",
    "dashscope>=1.25.0",
    "httpcore[socks]>=1.0.9",
    "httpx[socks]>=0.28.1",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langgraph==1.0.2",
    "langgraph-cli>=0.1.65",
    "loguru>=0.7.3",
    "pymupdf>=1.26.6",
    "pytest>=9.0.1",
    "pyyaml>=6.0.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.13",
]
